Added rtigo9_omm example.
Same as rtigo9 but using Opacity Micromaps (OMM) for cutout textures which require OptiX SDK 7.6.0.
The OMM generation is done with the OptiX Toolkit CUDA OmmBaking tool.
The example is only built when OptiX SDK 7.6.0 and OptiX Toolkit are found.
Added FindOptiXToolkit.cmake.
Updated README.md with the description of the rtigo9_omm example and the build in